+++ covalent_dispatcher/_core/data_manager.py
@@ -28,8 +28,9 @@ import uuid
 from covalent._results_manager import Result
 from covalent._shared_files import logger
 from covalent._workflow.lattice import Lattice
+from covalent._workflow.transport_graph_ops import TransportGraphOps
 
-from .._db import update, upsert
+from .._db import load, update, upsert
 from .._db.write_result_to_db import resolve_electron_id
 
 app_log = logger.app_log
@@ -142,6 +143,39 @@ def make_dispatch(
     )
     _register_result_object(result_object)
     return result_object.dispatch_id
+
+
+def make_derived_dispatch(
+    parent_dispatch_id: str, json_lattice: str, electron_updates=None, reuse_previous_results=False
+):
+    if electron_updates is None:
+        electron_updates = {}
+    old_res = load.get_result_object_from_storage(parent_dispatch_id, wait=False)
+
+    if json_lattice:
+        lat = Lattice.deserialize_from_json(json_lattice)
+        res = Result(lat, get_unique_id())
+        res._initialize_nodes()
+
+        tg = res.lattice.transport_graph
+        tg_old = old_res.lattice.transport_graph
+        if reuse_previous_results:
+            # reusable_nodes = TransportGraphDiffer.compare_transport_graphs(tg_old, tg)
+            reusable_nodes = TransportGraphOps(tg_old).get_diff_nodes(tg)
+            tg_old.copy_nodes(tg, reusable_nodes)
+            print(f"Reused nodes {reusable_nodes} from {parent_dispatch_id}")
+    else:
+        res = Result(old_res.lattice, get_unique_id())
+        res._num_nodes = old_res._num_nodes
+        if not reuse_previous_results:
+            res._initialize_nodes()
+
+    res.lattice.transport_graph.apply_electron_updates(electron_updates)
+    res.lattice.transport_graph.dirty_nodes = list(res.lattice.transport_graph._graph.nodes)
+    update.persist(res)
+    _register_result_object(res)
+
+    return res.dispatch_id
 
 
 def get_result_object(dispatch_id: str) -> Result: