+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Workflow re-dispatching functionality.
+- Transport graph diff'ing for workflow re-dispatching.
 
 ## [0.214.0-rc.0] - 2023-01-25